CheckPaymentDetails_BE/requirements.txt
asgiref>=3.9.1
certifi>=2025.8.3
charset-normalizer>=3.4.3
Django>=5.2.4
djangorestframework>=3.16.0
dotenv>=0.9.9
gunicorn>=23.0.0
idna>=3.10
numpy>=2.3.2
overrides>=7.7.0
pandas>=2.3.1
PyMySQL>=1.1.1
python-dateutil>=2.9.0.post0
python-dotenv>=1.1.1
pytz>=2025.2
RapidFuzz>=3.13.0
requests>=2.32.5
six>=1.17.0
sqlparse>=0.5.3
tzdata>=2025.2
urllib3>=2.5.0

mysqlclient>=2.2.7